split build.rs unifinshed
